requirements.txt
tensorflow>=2.12.0
tensorflow-model-optimization>=0.7.3
tf2onnx>=1.14.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
scikit-image>=0.20.0
SimpleITK>=2.2.0
nibabel>=5.0.0
Pillow>=9.4.0
opencv-python>=4.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.64.0
wandb>=0.14.0
kaggle>=1.5.13
